Extract edit image tracking to separate class
