+++ WordPress/src/main/java/org/wordpress/android/WordPress.java
@@ -77,6 +77,7 @@ import org.wordpress.android.ui.notifications.SystemNotificationsTracker;
 import org.wordpress.android.ui.notifications.services.NotificationsUpdateServiceStarter;
 import org.wordpress.android.ui.notifications.utils.NotificationsUtils;
 import org.wordpress.android.ui.posts.editor.ImageEditorInitializer;
+import org.wordpress.android.ui.posts.editor.ImageEditorTracker;
 import org.wordpress.android.ui.prefs.AppPrefs;
 import org.wordpress.android.ui.reader.tracker.ReaderTracker;
 import org.wordpress.android.ui.stats.refresh.lists.widget.WidgetUpdater.StatsWidgetUpdaters;
@@ -157,6 +158,7 @@ public class WordPress extends MultiDexApplication implements HasServiceInjector
     @Inject SystemNotificationsTracker mSystemNotificationsTracker;
     @Inject ReaderTracker mReaderTracker;
     @Inject ImageManager mImageManager;
+    @Inject ImageEditorTracker mImageEditorTracker;
 
     // For development and production `AnalyticsTrackerNosara`, for testing a mocked `Tracker` will be injected.
     @Inject Tracker mTracker;
@@ -324,7 +326,7 @@ public class WordPress extends MultiDexApplication implements HasServiceInjector
         UploadWorkerKt.enqueuePeriodicUploadWorkRequestForAllSites();
 
         mSystemNotificationsTracker.checkSystemNotificationsState();
-        ImageEditorInitializer.Companion.init(mImageManager);
+        ImageEditorInitializer.Companion.init(mImageManager, mImageEditorTracker);
     }
 
     protected void initWorkManager() {
